qa/qa_api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
asgiref==3.7.2
django-redis==5.4.0
redis==5.0.1
requests==2.32.0
GitPython==3.1.41
openpyxl==3.0.3
user_agents==2.2.0
attrs>=21.3.0
sqlparse==0.4.4
pyyaml==6.0.1